refactor: make properties class immutable
