+++ backend/src/main/java/com/doodledoodle/backend/config/audit/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.doodledoodle.backend.config;
+package com.doodledoodle.backend.config.audit;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
