+++ backend/src/main/java/com/doodledoodle/backend/config/properties/PropertiesConfig.java
@@ -1,0 +1,9 @@
+package com.doodledoodle.backend.config.properties;
+
+import com.doodledoodle.backend.global.storage.S3StorageProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties({S3StorageProperties.class})
+public class PropertiesConfig {}
