+++ backend/src/main/java/com/doodledoodle/backend/global/storage/S3StorageProperties.java
@@ -1,16 +1,18 @@
 package com.doodledoodle.backend.global.storage;
 
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.ConstructorBinding;
 import org.springframework.stereotype.Component;
 
 @Getter
-@Setter
-@Component
+@ConstructorBinding
+@RequiredArgsConstructor
 @ConfigurationProperties(prefix = "cloud.aws.s3")
 public class S3StorageProperties {
-    private String uri;
-    private String bucket;
+    private final String uri;
+    private final String bucket;
 }
